section 8: split hfs_fsck Phase 3 (catalog) into submodule, completing the split
Extract catalog consistency checks + repairs into
src/fs/hfs_fsck/catalog.rs (1147 lines): CatalogKey, CatalogEntry,
collect_catalog_entries, parse_catalog_record, check_directory_structure,
check_catalog_consistency, repair_catalog_consistency,
insert_thread_record, fix_thread_record, repair_thread_record_lengths,
remove_thread_record, remove_record_by_key, fix_valence_mismatches.

bitmap.rs now imports CatalogEntry / collect_catalog_entries from
super::catalog::... instead of from mod.rs. catalog.rs imports
validate_hfs_name from super::mdb::... and the thread-type constants
CATALOG_DIR_THREAD / CATALOG_FILE_THREAD from mod.rs (now pub(super)).

hfs_fsck/mod.rs drops from 3054 to 1916 lines — only orchestrators
(check_hfs_integrity, repair_hfs), the HfsFsckCode enum, and tests
remain. The original 6588-line monolith is now five phase-focused
submodules totaling 4779 + 1916 = 6695 lines (slight growth from
per-file headers and re-export plumbing). All 61 fsck tests pass.
