+++ docs/codecleanup.md
@@ -176,9 +176,14 @@ Conventions:
 - [ ] Largest area of recent change — confirm `hfs_common.rs` exposes everything `hfs_clone.rs` needs (`pub(crate)` items have crept in over time).
 - [ ] Catalog walking / record-range helpers shared with HFS+ (§1).
 - [ ] Audit debug `eprintln!` instances flagged in `hfs.rs:3068–3124` and replace with `log::debug!`.
-- [~] `hfs_fsck.rs` split started: extents-overflow code moved to its own submodule.
-  - **Done:** `src/fs/hfs_fsck.rs` is now `src/fs/hfs_fsck/mod.rs` (3054 lines, was 6588) plus four submodules: `extents.rs` (1335 lines, the 17 `*_extents_*` functions for the extents-overflow B-tree), `mdb.rs` (260 lines, Phase 1 MDB sanity), `bitmap.rs` (698 lines, Phase 4 extent/allocation-bitmap cross-check + repairs), and `btree.rs` (1322 lines, Phase 2 catalog B-tree structure checks + repairs — `check_btree_structure`, `check_map_nodes`, `check_key_ordering`, `walk_leaf_chain`, `check_index_sibling_links`, etc., plus repairs `repair_btree_structure_leaves_only`, `repair_map_nodes`, `repair_node_bitmap`, `repair_key_ordering`, `rebuild_leaf_chain`, `rebuild_index_nodes`). The shared btree primitives `record_key`, `extract_child_pointers`, `HFS_MAX_NRECS` moved with btree.rs and extents.rs imports them via `super::btree::...`. `rebuild_index_nodes` re-exported from mod.rs as `pub(crate)` so `hfs.rs` callers (`super::hfs_fsck::rebuild_index_nodes`) keep working. `CatalogKey`, `CatalogEntry`, `collect_catalog_entries`, `hfs_issue`, `HfsFsckCode` are `pub(super)` for cross-submodule access. All 61 fsck tests still pass under their new module paths.
-  - **Not yet done:** Phase 3 (catalog consistency) split — `check_directory_structure`, `check_catalog_consistency`, `parse_catalog_record`, plus the catalog repair fns (`repair_catalog_consistency`, `insert_thread_record`, `fix_thread_record`, `repair_thread_record_lengths`, `remove_thread_record`, `remove_record_by_key`, `fix_valence_mismatches`). With Phase 2 settled the catalog split should be the cleanest of the four — it can pull btree helpers from `super::btree::...` and just owns the catalog parser plus its repairs.
+- [x] `hfs_fsck.rs` fully split by phase.
+  - `src/fs/hfs_fsck.rs` (was 6588 lines) is now `src/fs/hfs_fsck/mod.rs` (1916 lines, just orchestrators + tests + the `HfsFsckCode` enum) plus five phase submodules:
+    - `mdb.rs` (260 lines) — Phase 1 MDB sanity.
+    - `btree.rs` (1338 lines) — Phase 2 catalog B-tree structure checks + repairs, plus the shared btree primitives `record_key`, `extract_child_pointers`, `HFS_MAX_NRECS`.
+    - `catalog.rs` (1147 lines) — Phase 3 catalog consistency checks + repairs (`CatalogKey`, `CatalogEntry`, `collect_catalog_entries`, `parse_catalog_record`, `check_directory_structure`, `check_catalog_consistency`, `repair_catalog_consistency`, `insert_thread_record`, `fix_thread_record`, `repair_thread_record_lengths`, `remove_thread_record`, `remove_record_by_key`, `fix_valence_mismatches`).
+    - `bitmap.rs` (699 lines) — Phase 4 extent / allocation-bitmap cross-check + repairs.
+    - `extents.rs` (1335 lines) — extents-overflow B-tree checks + repairs.
+  - Cross-module wiring: `extents.rs` and `catalog.rs` import shared btree primitives via `super::btree::...`; `bitmap.rs` imports `CatalogEntry` / `collect_catalog_entries` from `super::catalog::...`; `catalog.rs` imports `validate_hfs_name` from `super::mdb::...`. `rebuild_index_nodes` is re-exported from mod.rs as `pub(crate)` so `hfs.rs` callers (`super::hfs_fsck::rebuild_index_nodes`) keep working unchanged. `hfs_issue`, `HfsFsckCode`, `CATALOG_DIR_THREAD`, `CATALOG_FILE_THREAD` stay in mod.rs as `pub(super)` shared helpers. All 61 fsck tests still pass.
 
 ### HFS+ (`src/fs/hfsplus.rs`, 2607)
 - [ ] Migrate to `btree_split_leaf_with_insert` (§1).
